Add files via upload
A program for determining the 8 bit binary sequence for IP addresses. So for the IP address 154.31.16.13 the first numeric sequence is 154 and the next numeric sequence is 31, etc. So in this program, the IP address 154.31.16.13 would be: 10011010.00011111.00010000.00001101
In addition, the first and firstNum variables combined with the if/else clauses determine what class the IP address is. (Assuming it is an IPv4 IP address).