water use and string manipulation programs
Includes a string manipulations folder with a file that converts all letters to uppercase, counts and displays the frequency of each letter in the string, yet avoids counting or displaying non alphabetical characters.

Also includes a folder with water usage and billing programs. The water usage program writes data to a .txt file called water.txt (and assumes this file is in the same folder). The water billing program reads the data from the file and calculates the water bill given a set of condtions.